backend/pyproject.toml
[tool.poetry]
name = "coin-maker-backend"
version = "0.1.0"
description = "FastAPI backend for 3D Coin Maker - 3D printable coin generator"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
Pillow = "^10.1.0"
celery = "^5.3.4"
redis = "^5.0.1"
python-decouple = "^3.8"
numpy = "^1.24.0"
manifold3d = "^3.2.0"
trimesh = "^4.0.0"
apscheduler = "^3.11.0"
fastapi = "^0.116.1"
uvicorn = "^0.35.0"
python-multipart = "^0.0.20"
pydantic-settings = "^2.5.2"

[tool.poetry.group.dev.dependencies]
gunicorn = "^21.2.0"
pyright = "^1.1.300"
ruff = "^0.1.0"
pytest = "^7.4.0"
requests = "^2.31.0"
httpx = "^0.28.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py311"
line-length = 100
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by formatter
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.pyright]
include = [".", "apps", "core"]
exclude = ["**/migrations", "**/__pycache__"]
defineConstant = { DEBUG = true }
reportMissingImports = true
reportMissingTypeStubs = false
pythonVersion = "3.11"
pythonPlatform = "Linux"